fw_att_control split header
