+++ src/modules/fw_att_control/fw_att_control_main.cpp
@@ -41,285 +41,9 @@
  *
  */
 
-#include <px4_config.h>
-#include <px4_defines.h>
-#include <px4_tasks.h>
-#include <px4_posix.h>
+#include "FixedwingAttitudeControl.hpp"
 
-#include <drivers/drv_hrt.h>
-#include <ecl/attitude_fw/ecl_pitch_controller.h>
-#include <ecl/attitude_fw/ecl_roll_controller.h>
-#include <ecl/attitude_fw/ecl_wheel_controller.h>
-#include <ecl/attitude_fw/ecl_yaw_controller.h>
-#include <geo/geo.h>
-#include <mathlib/mathlib.h>
-#include <systemlib/param/param.h>
-#include <systemlib/perf_counter.h>
-#include <uORB/topics/actuator_controls.h>
-#include <uORB/topics/battery_status.h>
-#include <uORB/topics/control_state.h>
-#include <uORB/topics/manual_control_setpoint.h>
-#include <uORB/topics/parameter_update.h>
-#include <uORB/topics/vehicle_attitude_setpoint.h>
-#include <uORB/topics/vehicle_control_mode.h>
-#include <uORB/topics/vehicle_global_position.h>
-#include <uORB/topics/vehicle_land_detected.h>
-#include <uORB/topics/vehicle_rates_setpoint.h>
-#include <uORB/topics/vehicle_status.h>
-#include <uORB/uORB.h>
-#include <vtol_att_control/vtol_type.h>
-
-using matrix::Eulerf;
-using matrix::Quatf;
-
-/**
- * Fixedwing attitude control app start / stop handling function
- *
- * @ingroup apps
- */
 extern "C" __EXPORT int fw_att_control_main(int argc, char *argv[]);
-
-class FixedwingAttitudeControl
-{
-public:
-	/**
-	 * Constructor
-	 */
-	FixedwingAttitudeControl();
-
-	/**
-	 * Destructor, also kills the main task.
-	 */
-	~FixedwingAttitudeControl();
-
-	/**
-	 * Start the main task.
-	 *
-	 * @return	PX4_OK on success.
-	 */
-	int		start();
-
-	/**
-	 * Task status
-	 *
-	 * @return	true if the mainloop is running
-	 */
-	bool		task_running() { return _task_running; }
-
-private:
-
-	bool		_task_should_exit;		/**< if true, attitude control task should exit */
-	bool		_task_running;			/**< if true, task is running in its mainloop */
-	int		_control_task;			/**< task handle */
-
-	int		_att_sp_sub;			/**< vehicle attitude setpoint */
-	int		_battery_status_sub;		/**< battery status subscription */
-	int		_ctrl_state_sub;		/**< control state subscription */
-	int		_global_pos_sub;		/**< global position subscription */
-	int		_manual_sub;			/**< notification of manual control updates */
-	int		_params_sub;			/**< notification of parameter updates */
-	int		_vcontrol_mode_sub;		/**< vehicle status subscription */
-	int		_vehicle_land_detected_sub;	/**< vehicle land detected subscription */
-	int		_vehicle_status_sub;		/**< vehicle status subscription */
-
-	orb_advert_t	_rate_sp_pub;			/**< rate setpoint publication */
-	orb_advert_t	_attitude_sp_pub;		/**< attitude setpoint point */
-	orb_advert_t	_actuators_0_pub;		/**< actuator control group 0 setpoint */
-	orb_advert_t	_actuators_2_pub;		/**< actuator control group 1 setpoint (Airframe) */
-
-	orb_id_t _rates_sp_id;	// pointer to correct rates setpoint uORB metadata structure
-	orb_id_t _actuators_id;	// pointer to correct actuator controls0 uORB metadata structure
-	orb_id_t _attitude_setpoint_id;
-
-	struct actuator_controls_s			_actuators;		/**< actuator control inputs */
-	struct actuator_controls_s			_actuators_airframe;	/**< actuator control inputs */
-	struct battery_status_s				_battery_status;	/**< battery status */
-	struct control_state_s				_ctrl_state;	/**< control state */
-	struct manual_control_setpoint_s		_manual;		/**< r/c channel data */
-	struct vehicle_attitude_setpoint_s		_att_sp;		/**< vehicle attitude setpoint */
-	struct vehicle_control_mode_s			_vcontrol_mode;		/**< vehicle control mode */
-	struct vehicle_global_position_s		_global_pos;		/**< global position */
-	struct vehicle_land_detected_s			_vehicle_land_detected;	/**< vehicle land detected */
-	struct vehicle_rates_setpoint_s			_rates_sp;	/* attitude rates setpoint */
-	struct vehicle_status_s				_vehicle_status;	/**< vehicle status */
-
-	perf_counter_t	_loop_perf;			/**< loop performance counter */
-	perf_counter_t	_nonfinite_input_perf;		/**< performance counter for non finite input */
-	perf_counter_t	_nonfinite_output_perf;		/**< performance counter for non finite output */
-
-	bool		_debug;				/**< if set to true, print debug output */
-
-	float _flaps_applied;
-	float _flaperons_applied;
-
-	struct {
-		float p_tc;
-		float p_p;
-		float p_i;
-		float p_ff;
-		float p_rmax_pos;
-		float p_rmax_neg;
-		float p_integrator_max;
-
-		float r_tc;
-		float r_p;
-		float r_i;
-		float r_ff;
-		float r_integrator_max;
-		float r_rmax;
-
-		float y_p;
-		float y_i;
-		float y_ff;
-		float y_integrator_max;
-		float roll_to_yaw_ff;
-		float y_rmax;
-
-		bool w_en;
-		float w_p;
-		float w_i;
-		float w_ff;
-		float w_integrator_max;
-		float w_rmax;
-
-		float airspeed_min;
-		float airspeed_trim;
-		float airspeed_max;
-
-		float trim_roll;
-		float trim_pitch;
-		float trim_yaw;
-
-		float rollsp_offset_deg;		/**< Roll Setpoint Offset in deg */
-		float pitchsp_offset_deg;		/**< Pitch Setpoint Offset in deg */
-		float rollsp_offset_rad;		/**< Roll Setpoint Offset in rad */
-		float pitchsp_offset_rad;		/**< Pitch Setpoint Offset in rad */
-
-		float man_roll_max;				/**< Max Roll in rad */
-		float man_pitch_max;			/**< Max Pitch in rad */
-
-		float man_roll_scale;			/**< scale factor applied to roll actuator control in pure manual mode */
-		float man_pitch_scale;			/**< scale factor applied to pitch actuator control in pure manual mode */
-		float man_yaw_scale; 			/**< scale factor applied to yaw actuator control in pure manual mode */
-
-		float acro_max_x_rate_rad;
-		float acro_max_y_rate_rad;
-		float acro_max_z_rate_rad;
-
-		float flaps_scale;				/**< Scale factor for flaps */
-		float flaperon_scale;			/**< Scale factor for flaperons */
-
-		float rattitude_thres;
-
-		int32_t vtol_type;					/**< VTOL type: 0 = tailsitter, 1 = tiltrotor */
-
-		int32_t bat_scale_en;			/**< Battery scaling enabled */
-
-	}		_parameters{};			/**< local copies of interesting parameters */
-
-	struct {
-
-		param_t p_tc;
-		param_t p_p;
-		param_t p_i;
-		param_t p_ff;
-		param_t p_rmax_pos;
-		param_t p_rmax_neg;
-		param_t p_integrator_max;
-
-		param_t r_tc;
-		param_t r_p;
-		param_t r_i;
-		param_t r_ff;
-		param_t r_integrator_max;
-		param_t r_rmax;
-
-		param_t y_p;
-		param_t y_i;
-		param_t y_ff;
-		param_t y_integrator_max;
-		param_t roll_to_yaw_ff;
-		param_t y_rmax;
-
-		param_t w_en;
-		param_t w_p;
-		param_t w_i;
-		param_t w_ff;
-		param_t w_integrator_max;
-		param_t w_rmax;
-
-		param_t airspeed_min;
-		param_t airspeed_trim;
-		param_t airspeed_max;
-
-		param_t trim_roll;
-		param_t trim_pitch;
-		param_t trim_yaw;
-
-		param_t rollsp_offset_deg;
-		param_t pitchsp_offset_deg;
-
-		param_t man_roll_max;
-		param_t man_pitch_max;
-
-		param_t man_roll_scale;
-		param_t man_pitch_scale;
-		param_t man_yaw_scale;
-
-		param_t acro_max_x_rate;
-		param_t acro_max_y_rate;
-		param_t acro_max_z_rate;
-
-		param_t flaps_scale;
-		param_t flaperon_scale;
-
-		param_t rattitude_thres;
-
-		param_t vtol_type;
-
-		param_t bat_scale_en;
-
-	}		_parameter_handles{};		/**< handles for interesting parameters */
-
-	// Rotation matrix and euler angles to extract from control state
-	math::Matrix<3, 3> _R;
-	float _roll;
-	float _pitch;
-	float _yaw;
-
-	ECL_RollController				_roll_ctrl;
-	ECL_PitchController				_pitch_ctrl;
-	ECL_YawController				_yaw_ctrl;
-	ECL_WheelController			_wheel_ctrl;
-
-
-	/**
-	 * Update our local parameter cache.
-	 */
-	void		parameters_update();
-
-	/**
-	 * Check for changes in vehicle control mode.
-	 */
-	void		vehicle_control_mode_poll();
-	void		vehicle_manual_poll();
-	void		vehicle_setpoint_poll();
-	void		global_pos_poll();
-	void		vehicle_status_poll();
-	void		vehicle_land_detected_poll();
-	void		battery_status_poll();
-
-	/**
-	 * Shim for calling task_main from task_create.
-	 */
-	static void	task_main_trampoline(int argc, char *argv[]);
-
-	/**
-	 * Main attitude controller collection task.
-	 */
-	void		task_main();
-
-};
 
 namespace att_control
 {
@@ -841,6 +565,7 @@ FixedwingAttitudeControl::task_main()
 			/* Simple handling of failsafe: deploy parachute if failsafe is on */
 			if (_vcontrol_mode.flag_control_termination_enabled) {
 				_actuators_airframe.control[7] = 1.0f;
+
 			} else {
 				_actuators_airframe.control[7] = 0.0f;
 			}
